% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find.dest.control.r
\name{find.dest.control}
\alias{find.dest.control}
\title{Find destination control}
\usage{
find.dest.control(tu)
}
\arguments{
\item{tu}{turfID}
}
\value{
character vector with TT1 and TTC controls
}
\description{
Find turfID of controls in the destination block
}

