% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findfinal.r
\name{findfinal}
\alias{findfinal}
\title{Find target number of species}
\usage{
findfinal(extinct = TRUE)
}
\arguments{
\item{extinct}{extinct or immigrant}
}
\description{
Number of species needed to immigrate/go extinct to match target control
}

