% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findfinaltrait.r
\name{findfinaltrait}
\alias{findfinaltrait}
\title{Find target cover of trait}
\usage{
findfinaltrait(usecover = TRUE, keep)
}
\arguments{
\item{usecover}{Use cover or no subplots.}

\item{keep}{logical vector of which species to include}
}
\description{
Find cover of species with trait in 2013 in destination control
}

