% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.dest.control.community.r
\name{get.dest.control.community}
\alias{get.dest.control.community}
\title{Get destination control community}
\usage{
get.dest.control.community(tu, TT1 = TRUE, TTC = TRUE, year = 2009:2013)
}
\arguments{
\item{tu}{turfID}

\item{TT1}{get species list of TT1 turf}

\item{TTC}{get species list of TTC turf}

\item{year}{which years to use}
}
\description{
Get the species list of the destination control community
}

