% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.turf.community.r
\name{get.turf.community}
\alias{get.turf.community}
\title{Get Turf Community}
\usage{
get.turf.community(tu, year = 2009:2013)
}
\arguments{
\item{tu}{TurfID}

\item{year}{years to include}
}
\value{
character vector of species codes. "MISSING" is returned if no species are found (for example in a destroyed turf)
}
\description{
Get names of species in turf community in given year(s)
}

