% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hogsete.plot.r
\name{hogsete.plot}
\alias{hogsete.plot}
\title{Plots NMDS of transpants, origin and destinations}
\usage{
hogsete.plot(site, dat = cover, ord = metaMDS, ...)
}
\arguments{
\item{site}{which site}

\item{dat}{data to use - cover, subturf or presence absence}

\item{ord}{ordination methods to use  - should work with any from the vegan package.}

\item{\dots}{other arguments to ord (cannot think of anything useful)}
}
\description{
Plots NMDS of transpants, origin and destinations
}
\details{
This plot tends to be less ugly than the reverse hogsete plot.
}

