% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_community_data.r
\name{load_community_data}
\alias{load_community_data}
\title{Imports Community data from source files}
\usage{
load_community_data(db = "D:/downloadeddata/seedclim_2014-12-19.mdb",
  isDNS = FALSE, moss = FALSE)
}
\arguments{
\item{db}{The path to the database or DNS}

\item{isDNS}{Is db a DNS?}

\item{moss}{Load bryophyte data?}
}
\value{
Loads data into memory. No return value
}
\description{
Runs source files to load community data into memory.
}
\details{
Uses RODBC so only works with 32 bit version of windows.
Currently loads
\itemize{
  \item{Cover data}
  \item{Subturf frequency data}
  \item{Trait data}
  \item{Turf level environmet data}
  \item{Bryophyte subturf frequency data (optional)}
  \item{seedClim colours and symbols (not yet finalised)}
}
}
\examples{
\dontrun{load_community_data()}
}

