% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.directional.dists.r
\name{make.directional.dists}
\alias{make.directional.dists}
\title{Directional Taxomonic Distance Moved By Turfs}
\usage{
make.directional.dists(X)
}
\arguments{
\item{X}{data.frame of cover or subturf frequencies}
}
\value{
matrix of distances travelled by each turf
\itemize{
 \item{d11 distance travelled towards target between 2009 and 2011}
 \item{d12 distance travelled towards target between 2009 and 2012}
 \item{d13 distance travelled towards target between 2009 and 2013}
 \item{r11 relative distance travelled towards target between 2009 and 2011}
 \item{r12 relative distance travelled towards target between 2009 and 2012}
 \item{r13 relative distance travelled towards target between 2009 and 2013}
}
}
\description{
Finds Bray-Curtis distance moved by turf from its between original community towards its target community.
Uses destination TT1 turf as the target. If TT1 turf is missing, no value is given.
Relative distance show distance as a proportion of the original distance between turf community and the destination TT1 turf.
The all distances are to the TT1 community in the relevant year - i.e. to the current target, not the initial target.
}

