% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.dists.r
\name{make.dists}
\alias{make.dists}
\title{Taxomonic distance moved by turfs}
\usage{
make.dists(x)
}
\arguments{
\item{x}{data.frame of cover or subturf frequencies}
}
\value{
matrix of distances travelled by each turf
\itemize{
 \item{d911 distance travelled between 2009 and 2011}
 \item{d912 distance travelled between 2009 and 2012}
 \item{d913 distance travelled between 2009 and 2013}
}
}
\description{
Finds Bray-Curtis distance between original community and the community in subsequent years.
}

