% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multitrait.r
\name{multitraits}
\alias{multitraits}
\title{plot risk of extinction for different treatments}
\usage{
multitraits(trait, char = NULL, numeric, type = "extinction")
}
\arguments{
\item{trait}{which trait}

\item{char}{which value of character traits}

\item{numeric}{is the trait numeric?}

\item{type}{which contrast to show: extinction, immigrant or persistant}
}
\value{
multitrait object
}
\description{
plot risk of extinction for different treatments
}

