% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numberspecies.r
\name{numberspecies}
\alias{numberspecies}
\title{Main winners and loosers plot}
\usage{
numberspecies(x, ylab = NULL, main = NULL, show = "p", final,
  years = 2:4, above = TRUE, inarow = FALSE, bottomrow = FALSE)
}
\arguments{
\item{x}{result from winlose or trait selector (processed)}

\item{ylab}{y-axis label}

\item{main}{title}

\item{show}{turfs aggregated by "p" preciptiation and or "t" temperature levels}

\item{final}{result from findfinal or findfinaltrait}

\item{years}{which year are represented by the data}

\item{above}{plot the title above or below the plot border}

\item{inarow}{are the plots all in a row (gives yaxis labels only on first plot)}

\item{bottomrow}{if inarow, is this the bottomrow? If so, give axis labels etc.}
}
\description{
Main winners and loosers plot
}

