% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.towards.target.r
\name{plot.towards.target}
\alias{plot.towards.target}
\title{Boxplots of directional distance moved}
\usage{
\method{plot}{towards.target}(x, split = TRUE, temperature = TRUE,
  draw = 2011:2013, notch = TRUE, addmain = TRUE, setpar = TRUE, ...)
}
\arguments{
\item{x}{output of directional distance moved cbind() to turfs}

\item{split}{Split the data by temperature/precipitation?}

\item{temperature}{If split == TRUE, make the split by temperature or precipitation}

\item{draw}{which years to plot}

\item{notch}{logical add notches}

\item{addmain}{add main title to plot}

\item{setpar}{let function set graphical parameters}

\item{\dots}{currently unused}
}
\description{
Boxplots of directional distance moved
}

