% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotWL.r
\name{plotWL}
\alias{plotWL}
\title{plot of winners and losers}
\usage{
plotWL(W, ylim)
}
\arguments{
\item{W}{winners and losers}

\item{ylim}{y axis limits}
}
\description{
plot of winners and losers
}

