% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trait.extinction.plot.r
\name{trait.extinction.plot}
\alias{trait.extinction.plot}
\title{trait extinction plot}
\usage{
trait.extinction.plot(sites, treatment, trait, numeric, char = NULL,
  plot = TRUE, type = "extinction")
}
\arguments{
\item{sites}{which sites to include}

\item{treatment}{which treatment to include}

\item{trait}{which trait to include}

\item{numeric}{is the trait numeric}

\item{char}{value of character trait}

\item{plot}{should the result be plotted/}

\item{type}{which contrast should be used: extinction, immigant or persistant}
}
\value{
result that can be used by multitrait
}
\description{
trait extinction plot
}

