% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/winlose.r
\name{winlose}
\alias{winlose}
\title{Which taxa are winning and losing}
\usage{
winlose(turf, year1, year2, rare = 4, relative = 0.2)
}
\arguments{
\item{turf}{turfID}

\item{year1}{start year of comparison}

\item{year2}{end year of comparison}

\item{rare}{what constitutes a rare species}

\item{relative}{proportion change needed not to be considered a generalist}
}
\description{
best used in a loop
}

